Remove rowSums, colSums, rowMeans, colMeans, mean, and t generics
The implicit generic for each of these functions (i.e. the S4 generic that
gets automatically created the first time setMethod() is used to define a
method for the function) is exactly how it should be, that is, it has the
right arguments and signature. So there is really no need for us to expli-
citely define these generics upfront in BiocGenerics with setGeneric()
statements that result in exactly the same generics as these implicit
generics.