Revert "Remove rowSums, colSums, rowMeans, colMeans, mean, and t generics"
This reverts commit 6e2a09790a15bc872f5682a9dafd9eb8ce20a40c.

Even though the implicit generics for these functions are exactly how they
should be, Bioconductor packages that call these functions (with the hope
that they will be able to dispatch to the appropriate methods) still need
to be able to import the generics from "somewhere". Explicitly defining
these generics in BiocGenerics addresses that.

Also simplify setGeneric statements for row/colSums and row/colMeans.